rework: Results are being fetched in the wrong location
